requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
banks==2.2.0
beautifulsoup4==4.14.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.3
cryptography==46.0.3
dataclasses-json==0.6.7
datapizza-ai==0.0.9
datapizza-ai-clients-openai==0.0.9
datapizza-ai-core==0.0.13
datapizza-ai-embedders-openai==0.0.5
datapizza-ai-parsers-docling==0.0.8
datapizza-ai-vectorstores-qdrant==0.0.5
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
dirtyjson==1.0.8
distro==1.9.0
docling==2.63.0
docling-core==2.54.0
docling-ibm-models==3.10.2
docling-parse==4.7.1
easyocr==1.7.2
et_xmlfile==2.0.0
executing==2.2.1
Faker==38.2.0
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
greenlet==3.2.4
griffe==1.15.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
latex2mathml==3.78.1
lazy_loader==0.4
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.8
llama-index-cli==0.5.3
llama-index-core==0.14.8
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-openai==0.6.10
llama-index-readers-file==0.5.5
llama-index-readers-llama-parse==0.5.1
llama-index-readers-pdf-marker==0.4.1
llama-index-workflows==2.11.5
llama-parse==0.6.54
lxml==6.0.2
markdown-it-py==4.0.0
marker-pdf==0.2.17
marko==2.2.1
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
mcp==1.22.0
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6
ninja==1.13.0
nltk==3.9.2
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.23.2
openai==2.8.1
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pandas==2.2.3
parso==0.8.5
pdftext==0.3.20
pillow==10.4.0
platformdirs==4.5.0
pluggy==1.6.0
polyfactory==3.1.0
portalocker==3.2.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pure_eval==0.2.3
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pylatexenc==2.10
PyMuPDF==1.26.6
pypdf==6.4.0
pypdfium2==4.30.0
pyreadline3==3.5.4
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.16.1
RapidFuzz==3.14.3
rapidocr==3.4.2
referencing==0.37.0
regex==2024.11.6
requests==2.32.5
rich==14.2.0
rpds-py==0.29.0
rtree==1.4.1
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.4.2
scipy==1.16.3
semchunk==2.2.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.50.0
striprtf==0.0.26
surya-ocr==0.5.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
test-tecnico-ai-engineer @ file:///C:/Users/g.liturri/Desktop/data-pizza/test-tecnico-ai-engineer
texify==0.1.10
threadpoolctl==3.6.0
tifffile==2025.10.16
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
wrapt==1.17.3
xai-sdk==1.5.0
xlsxwriter==3.2.9
yarl==1.22.0
zipp==3.23.0
